requirements.txt
anthropic>=0.25.0
openai>=1.30.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
pytest>=8.0.0
